fix(logic): clear pre-existing -D warnings debt surfaced by the make check/clippy fix
Now that `cargo make check`/`cargo make clippy` actually reach bcinr-logic
with the real --workspace --all-targets --all-features -D warnings
invocation (see the companion Makefile.toml fix), two layers of
pre-existing lint debt were exposed:

1. Three rustc-level (non-clippy) issues under `cargo check --all-targets
   --all-features -D warnings` that were blocking the workspace-wide
   check entirely:
   - sigmoid_sat_u32.rs: unnecessary_parens in a test literal.
   - modular_mul_u64.rs: an unused local `const M` shadowed by (and
     redundant with) an inner helper's own `const M` -- deleted, the
     helper's own constant already does the real work.
   - reverse_bits_u128.rs: three NEGATIVE MUTANT functions
     (mutant_reverse_bits_u128_1/2/3) declared but never wired into an
     assertion -- dead code, and more importantly untested coverage for
     the bug classes they represent. Wired them into the same "mutant
     divergence" assert_ne! pattern already established and proven live
     in sibling algorithm files (e.g. bit_swap_u64.rs), rather than
     inventing a new convention.

2. The 14 pre-existing `cargo clippy -p bcinr-logic --lib` errors named in
   this session's gap-hunt report (identity_op, get_first,
   unnecessary_cast x3, manual_swap, double_must_use, let_unit_value x3,
   assign_op_pattern, manual_range_contains x3), plus two follow-on lints
   clippy could not previously reach because the first 14 were hard
   compile errors under -D warnings (manual_is_ascii_check on the newly
   rewritten range check, and rustc's path_statements on
   clippy::let_unit_value's own suggested fix for the hierarchical time
   wheel's compile-time assert! trick -- kept the `let _ = ...` form with
   an explicit, justified #[allow(clippy::let_unit_value)] instead of
   following that suggestion literally, since a bare path-statement there
   compiles and evaluates the const but rustc's own lint then (correctly,
   from its local view) calls it a no-op).

Every fix here is either a verified-equivalent algebraic simplification
(e.g. `u64::MAX & sat_mask` is provably always equal to `sat_mask` since
sat_mask is exactly 0 or all-ones by construction) or a pure syntax
rewrite with no semantic change (API rename, operator-assignment form,
range-contains, redundant-cast removal, redundant-attribute removal).
No behavior was changed in any safety-critical branchless algorithm.

Verified: cargo test -p bcinr-logic --lib (825 tests, unchanged pass
count) and cargo clippy -p bcinr-logic --lib both pass clean under
-D warnings; cargo fmt -p bcinr-logic made no further changes.

NOT closed by this commit (reported honestly, out of this session's
scope): `cargo clippy -p bcinr-logic --all-targets --all-features`
surfaces roughly 55 further pre-existing lint errors (type_complexity,
absurd_extreme_comparisons, approx_constant, byte_char_slices,
manual_checked_ops, and others) in test/bench targets this session did
not audit -- a substantially larger and previously-uncatalogued surface
than the 14 --lib-scoped errors this commit closes, discovered only
because the Makefile.toml fix now actually drives clippy against the
real --all-targets --all-features flag set for the first time.

+++ crates/bcinr-logic/src/fix.rs
@@ -346,7 +346,7 @@ pub fn q16_sin_approx(theta: i32) -> i32 {
     let t = {
         let raw = theta % TWO_PI;
         // Branchless: if raw < 0, add TWO_PI.
-        let neg_mask = (raw >> 31) as i32; // -1 if negative, 0 otherwise
+        let neg_mask = raw >> 31; // -1 if negative, 0 otherwise
         raw + (TWO_PI & neg_mask)
     };
     // Determine if t is in [pi, 2*pi]: second_half = 1 if t >= PI, else 0.
